refactored tests for parsers in input pkg;
