+++ src/test/software/amazon/event/ruler/input/InputWildcardTest.java
@@ -3,7 +3,7 @@ package software.amazon.event.ruler.input;
 import org.junit.Test;
 
 import static software.amazon.event.ruler.input.InputCharacterType.WILDCARD;
-import static software.amazon.event.ruler.input.Parser.ASTERISK_BYTE;
+import static software.amazon.event.ruler.input.DefaultParser.ASTERISK_BYTE;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
